requirements.txt
# Core
pyspark==3.5.1
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# App
streamlit>=1.36.0
plotly>=5.20.0

# Dev
tqdm>=4.66.0
pytest>=8.0.0
python-dotenv>=1.0.1
PyYAML>=6.0.1

